rename rules::requirement to rules::req_ref
